requirements.txt
numpy<1.17
pandas
python-telegram-bot
tensorflow
pickle-mixin
opencv-python-headless

